Test some run stuff.
