+++ crates/action-graph/tests/action_graph_test.rs
@@ -1133,444 +1133,6 @@ mod action_graph {
         }
     }
 
-    mod run_task_dependencies {
-        use super::*;
-
-        #[tokio::test]
-        async fn runs_deps_in_parallel() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let project = container.workspace_graph.get_project("deps").unwrap();
-            let task = container
-                .workspace_graph
-                .get_task_from_project(&project.id, "parallel")
-                .unwrap();
-
-            builder
-                .run_task(&project, &task, &RunRequirements::default())
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn runs_deps_in_serial() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let project = container.workspace_graph.get_project("deps").unwrap();
-            let task = container
-                .workspace_graph
-                .get_task_from_project(&project.id, "serial")
-                .unwrap();
-
-            builder
-                .run_task(&project, &task, &RunRequirements::default())
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn can_create_a_chain() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let project = container.workspace_graph.get_project("deps").unwrap();
-            let task = container
-                .workspace_graph
-                .get_task_from_project(&project.id, "chain1")
-                .unwrap();
-
-            builder
-                .run_task(&project, &task, &RunRequirements::default())
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn doesnt_include_dependents() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let project = container.workspace_graph.get_project("deps").unwrap();
-            let task = container
-                .workspace_graph
-                .get_task_from_project(&project.id, "base")
-                .unwrap();
-
-            builder
-                .run_task(&project, &task, &RunRequirements::default())
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn includes_dependents() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let project = container.workspace_graph.get_project("deps").unwrap();
-            let task = container
-                .workspace_graph
-                .get_task_from_project(&project.id, "base")
-                .unwrap();
-
-            builder
-                .run_task(
-                    &project,
-                    &task,
-                    &RunRequirements {
-                        dependents: true,
-                        ..RunRequirements::default()
-                    },
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn includes_dependents_for_ci() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let project = container.workspace_graph.get_project("deps").unwrap();
-            let task = container
-                .workspace_graph
-                .get_task_from_project(&project.id, "base")
-                .unwrap();
-
-            builder
-                .run_task(
-                    &project,
-                    &task,
-                    &RunRequirements {
-                        ci: true,
-                        dependents: true,
-                        ..RunRequirements::default()
-                    },
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-    }
-
-    mod run_task_by_target {
-        use super::*;
-
-        #[tokio::test]
-        #[should_panic(expected = "Dependencies scope (^:) is not supported in run contexts.")]
-        async fn errors_on_parent_scope() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("^:build").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-        }
-
-        #[tokio::test]
-        #[should_panic(expected = "Self scope (~:) is not supported in run contexts.")]
-        async fn errors_on_self_scope() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("~:build").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-        }
-
-        #[tokio::test]
-        async fn runs_all() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse(":build").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn runs_all_with_query() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder.set_query("language=rust").unwrap();
-
-            builder
-                .run_task_by_target(
-                    Target::parse(":build").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn runs_all_no_nodes() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse(":unknown").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert!(graph.is_empty());
-        }
-
-        #[tokio::test]
-        async fn doesnt_run_all_internal() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse(":internal").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert!(graph.is_empty());
-        }
-
-        #[tokio::test]
-        async fn runs_by_project() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("client:lint").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        #[should_panic(
-            expected = "No project has been configured with the identifier or alias unknown."
-        )]
-        async fn errors_for_unknown_project() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("unknown:build").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-        }
-
-        #[tokio::test]
-        #[should_panic(expected = "Unknown task unknown for project server.")]
-        async fn errors_for_unknown_project_task() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("server:unknown").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-        }
-
-        #[tokio::test]
-        #[should_panic(expected = "Unknown task internal for project common.")]
-        async fn errors_for_internal_task_when_explicit() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let locator = TargetLocator::Qualified(Target::parse("common:internal").unwrap());
-
-            builder
-                .run_task_by_target(
-                    Target::parse("common:internal").unwrap(),
-                    &RunRequirements {
-                        target_locators: FxHashSet::from_iter([locator]),
-                        ..RunRequirements::default()
-                    },
-                )
-                .unwrap();
-        }
-
-        #[tokio::test]
-        #[should_panic(expected = "Unknown task internal for project common.")]
-        async fn errors_for_internal_task_when_explicit_via_dir() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let locator = TargetLocator::TaskFromWorkingDir(Id::raw("internal"));
-
-            builder
-                .run_task_by_target(
-                    Target::parse("common:internal").unwrap(),
-                    &RunRequirements {
-                        target_locators: FxHashSet::from_iter([locator]),
-                        ..RunRequirements::default()
-                    },
-                )
-                .unwrap();
-        }
-
-        #[tokio::test]
-        async fn doesnt_error_for_internal_task_when_implicit() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("common:internal").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn doesnt_error_for_internal_task_when_depended_on() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            let locator = TargetLocator::Qualified(Target::parse("common:internal").unwrap());
-
-            builder
-                .run_task_by_target(
-                    Target::parse("misc:requiresInternal").unwrap(),
-                    &RunRequirements {
-                        target_locators: FxHashSet::from_iter([locator]),
-                        ..RunRequirements::default()
-                    },
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn runs_tag() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("#frontend:lint").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert_snapshot!(graph.to_dot());
-        }
-
-        #[tokio::test]
-        async fn runs_tag_no_nodes() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("#unknown:lint").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert!(graph.is_empty());
-        }
-
-        #[tokio::test]
-        async fn doesnt_run_tags_internal() {
-            let sandbox = create_sandbox("tasks");
-            let container = ActionGraphContainer::new(sandbox.path()).await;
-            let mut builder = container.create_builder();
-
-            builder
-                .run_task_by_target(
-                    Target::parse("#frontend:internal").unwrap(),
-                    &RunRequirements::default(),
-                )
-                .unwrap();
-
-            let graph = builder.build();
-
-            assert!(graph.is_empty());
-        }
-    }
-
     mod run_from_requirements {
         use super::*;
